requirements.txt
flask==3.1.0
opencv-python==4.10.0.84
numpy==2.2.1
pytesseract==0.3.13
pillow==11.0.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
werkzeug==3.1.3